fix: suggested changes done
