+++ package.json
@@ -1,9 +1,12 @@
 {
   "dependencies": {
     "@faker-js/faker": "^9.9.0",
+    "@types/bcrypt": "^6.0.0",
+    "@types/cookie-parser": "^1.4.9",
     "@types/express": "^5.0.3",
     "@types/lodash": "^4.17.20",
     "@types/request-ip": "^0.0.41",
+    "bcrypt": "^6.0.0",
     "cookie-parser": "^1.4.7",
     "dotenv": "^17.2.0",
     "express": "^5.1.0",
